Update Next-formalizationFull.hs
Average duration for events which change the ownership of the subject is added.
To handle agreements, I updated OEMEvent.oemEvtSubject characteristic as optional using Maybe. So far I used paymentRequired as a separator between invoice and delivery. But, here I see it is set as Always for a delivery also. If we say paymentRequired=Always for a delivery then there will be no difference between it and an invoice. How shall we continue?